Overhaul the timing framework to improve the global timer resolution to 1 microsecond and simplify the code, also for a later AArch64 port.
* Rename the udelay() syscall to sys_usleep() for consistence and accept a u64 parameter.
  udelay() with an u32 parameter is still provided to serve the existing customers, but should eventually vanish.

* Use time values in 1 microsecond granularity and u64 everywhere instead of simulating a 10ms timer with the CPU Time-Stamp Counter.
  This guarantees maximum precision for the best timing function we currently provide (sys_usleep).
  It also simplifies the code, because we can simply add microseconds to the timer tick count.

* Rewrite update_timer_ticks() as get_timer_ticks().
  It simply divides get_timestamp() by get_frequency() now to simulate a 1 microsecond timer with the CPU Time-Stamp Counter.
  This requires no per-core variables and is much more accurate.

* Calibrate the Local APIC Timer for 1 microsecond resolution.
  This reduces the maximum timeout to 34 seconds on a Intel Xeon E5-2650 v3 @ 2.30GHz, but for longer timeouts, the one-shot timer would simply
  fire multiple times.

* Detect CPU support for the TSC-Deadline Mode of the Local APIC Timer and use it.
  This one is easier to program than the legacy One-Shot Mode, even more accurate, and has no maximum timeout.

+++ librs/src/synch/semaphore.rs
@@ -87,7 +87,7 @@ impl Semaphore {
 	///
 	/// This method will block until the internal count of the semaphore is at
 	/// least 1.
-	pub fn acquire(&self, wakeup_time: Option<usize>) -> bool {
+	pub fn acquire(&self, wakeup_time: Option<u64>) -> bool {
 		// Reset last_wakeup_reason.
 		let core_scheduler = core_scheduler();
 		core_scheduler.current_task.borrow_mut().last_wakeup_reason = WakeupReason::Custom;
